create DigitalAssetConversion component & update landing footer & adjust files structure
