+++ src/components/ground_star_chain/digital_asset_conversion.tsx
@@ -1,0 +1,21 @@
+const DigitalAssetConversion = () => {
+  return (
+    <section className="relative mx-auto flex min-h-475px max-w-1440px items-center justify-start overflow-hidden">
+      <div className="absolute left-0 top-0 min-h-464px w-full translate-y-22% scale-200 bg-blueMoney bg-contain bg-top bg-no-repeat"></div>
+
+      <div className="relative w-1/2 space-y-12 px-80px">
+        <h2 className="text-48px font-semibold text-text-neutral-primary">
+          <span className="text-text-brand-primary-lv2">Digital Asset </span>Conversion
+        </h2>
+        <p className="text-lg font-medium text-text-neutral-primary">
+          Through the installation of iSunCoin nodes, clean energy generated by solar power systems
+          is transformed into digital assets, namely iSunCoin. These iSunCoins can be traded on the
+          blockchain network or used to pay for computational services on the platform, thus
+          digitizing and facilitating the circulation of energy value.
+        </p>
+      </div>
+    </section>
+  );
+};
+
+export default DigitalAssetConversion;
