+++ src/components/page_body/landing_page_body.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import HeroSectionOfLandingPage from '@/components/hero_section/hero_section_of_landing_page';
-import Introduction from '@/components/introduction/introduction';
-import StatisticsOverview from '@/components/statistics_overview/statistics_overview';
+import HeroSectionOfLandingPage from '@/components/landing_page_index/hero_section_of_landing_page';
+import Introduction from '@/components/landing_page_index/introduction';
+import StatisticsOverview from '@/components/landing_page_index/statistics_overview';
 
 const LandingPageBody = () => {
   return (
