+++ src/components/page_body/products_page_body.tsx
@@ -1,7 +1,7 @@
-import HeroSectionOfProducts from '@/components/hero_section/hero_section_of_products';
-import ThreeModelsIntro from '@/components/three_models/three_models_intro';
-import ThreeModelsVideo from '@/components/three_models/three_models_video';
-import ColorTypeSelector from '@/components/three_models/color_type_selector';
+import HeroSectionOfProducts from '@/components/products/hero_section_of_products';
+import ThreeModelsIntro from '@/components/products/three_models_intro';
+import ThreeModelsVideo from '@/components/products/three_models_video';
+import ColorTypeSelector from '@/components/products/color_type_selector';
 import ProductComparison from '@/components/products/product_comparison';
 import CustomerGuidance from '@/components/products/customer_guidance';
